requirements.txt
mediapipe
opencv-python
numpy
pystray
Pillow
